[Mosaic GPU] Cast i4->bf16 through i32
The cast was only registered in ml_dtypes very recently and we can't rely on it yet.

PiperOrigin-RevId: 810064672

+++ tests/mosaic/gpu_test.py
@@ -709,7 +709,7 @@ class WGMMALayoutTest(TestCase):
         low=sample_iinfo.min, high=sample_iinfo.max, size=(m, n), dtype=np.int32
     ).astype(jax_dtype_from)
 
-    expected = values.astype(jax_dtype_to)
+    expected = values.astype(np.int32).astype(jax_dtype_to)
     res = mgpu.as_gpu_kernel(
         kernel, (1, 1, 1), (128, 1, 1), values, expected, ()
     )(values)
